# swagger.json
{
   "openapi": "3.0.0",
   "info": {
      "version": "1.0.0",
      "title": "Store Manager",
      "description": "API docs for Store Manager.",
      "contact": {
         "name": "Emmanuel Raji",
         "email": "manuelraji15@gmail.com"
      }
   },
   "servers": [
      {
         "url": "https://onlinestoremanager.herokuapp.com/api/v1"
      }
   ],
   "components": {
      "securitySchemes": {
         "bearerAuth": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
         }
      }
   },
   "paths": {
...
      "/sales": {
         "get": {
            "tags": [
               "Sales"
            ],
            "security": [
               {
                  "bearerAuth": []
               }
            ],
            "summary": "gets all sale records",
            "responses": {
               "200": {
                  "description": "gets all the sale records in the store. route is accessible to only the admin record'",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "message": {
                                 "type": "string"
                              },
                              "code": {
                                 "type": "integer"
                              }
                           }
                        }
                     }
                  }
               },
               "403": {
                  "description": "Invalid token",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "error": {
                                 "type": "string"
                              },
                              "code": {
                                 "type": "integer"
                              }
                           }
                        }
                     }
                  }
               }
            }
         },
         "post": {
            "tags": [
               "Sales"
            ],
            "security": [
               {
                  "bearerAuth": []
               }
            ],
            "summary": "creates a new sale record",
            "requestBody": {
               "required": true,
               "content": {
                  "application/json": {
                     "schema": {
                        "type": "object"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "Successfully created order",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "message": {
                                 "type": "string"
                              },
                              "code": {
                                 "type": "integer"
                              }
                           }
                        }
                     }
                  }
               },
               "400": {
                  "description": "your cart is empty",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "error": {
                                 "type": "string"
                              },
                              "code": {
                                 "type": "integer"
                              }
                           }
                        }
                     }
                  }
               },
               "403": {
                  "description": "you are not authorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "error": {
                                 "type": "string"
                              },
                              "code": {
                                 "type": "integer"
                              }
                           }
                        }
                     }
                  }
               },
               "500": {
                  "description": "internal server error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "error": {
                                 "type": "string"
                              },
                              "code": {
                                 "type": "integer"
                              }
                           }
                        }
                     }
                  }
               }
            }
         }
      },
      "/sales/{saleId}": {
         "get": {
            "tags": [
               "Sales"
            ],
            "security": [
               {
                  "bearerAuth": []
               }
            ],
            "summary": "Returns a sale order",
            "parameters": [
               {
                  "name": "saleId",
                  "in": "path",
                  "required": true,
                  "description": "Enter id of the sale record order",
                  "schema": {
                     "type": "integer"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "Successful fetched sale record",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "array",
                           "items": {
                              "type": "object",
                              "properties": {
                                 "order_id": {
                                    "type": "integer"
                                 },
                                 "date": {
                                    "type": "integer"
                                 },
                                 "product_name": {
                                    "type": "string"
                                 },
                                 "product_price": {
                                    "type": "integer"
                                 },
                                 "product_quantity": {
                                    "type": "integer"
                                 },
                                 "total_price": {
                                    "type": "integer"
                                 },
                                 "user_id": {
                                    "type": "integer"
                                 },
                                 "user_name": {
                                    "type": "string"
                                 }
                              }
                           }
                        }
                     }
                  }
               },
               "403": {
                  "description": "you are not authorized",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "error": {
                                 "type": "string"
                              },
                              "code": {
                                 "type": "integer"
                              }
                           }
                        }
                     }
                  }
               },
               "404": {
                  "description": "order not found",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "error": {
                                 "type": "string"
                              },
                              "code": {
                                 "type": "integer"
                              }
                           }
                        }
                     }
                  }
               },
               "500": {
                  "description": "internal server error",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "error": {
                                 "type": "string"
                              },
                              "code": {
                                 "type": "integer"
                              }
                           }
                        }
                     }
                  }
               }
            }
         }
      }
   }
}